use u64

+++ l1-contracts/contracts/state-transition/chain-interfaces/IExecutor.sol
@@ -148,9 +148,9 @@ interface IExecutor is IZKChainBase {
         L2DACommitmentScheme daCommitmentScheme;
         bytes32 daCommitment;
         uint64 firstBlockTimestamp;
-        uint256 firstBlockNumber;
+        uint64 firstBlockNumber;
         uint64 lastBlockTimestamp;
-        uint256 lastBlockNumber;
+        uint64 lastBlockNumber;
         uint256 chainId;
         bytes operatorDAInput;
     }
@@ -270,8 +270,8 @@ interface IExecutor is IZKChainBase {
     /// @notice Reports the block range for a zksync os batch.
     /// @dev IMPORTANT: in this release this range is not trusted and provided by the operator.
     event ReportCommittedBatchRangeZKsyncOS(
-        uint256 indexed batchNumber,
-        uint256 indexed firstBlockNumber,
-        uint256 indexed lastBlockNumber
+        uint64 indexed batchNumber,
+        uint64 indexed firstBlockNumber,
+        uint64 indexed lastBlockNumber
     );
 }
